+++ constants/directories.py
@@ -1,0 +1,4 @@
+# constants/directories.py
+
+MODELS_DIR_PATH = "models"
+VAE_DIR_PATH = "vae"
