+++ constants/schedulers.py
@@ -1,0 +1,8 @@
+# constants/schedulers.py
+
+SCHEDULERS = [
+    'DDIMScheduler',
+    'LMSDiscreteScheduler',
+    'PNDMScheduler',
+    'UniPCMultistepScheduler'
+]
