+++ predict.py
@@ -1,49 +1,15 @@
+# predict.py
+
 from cog import BasePredictor, Input, Path
-import diffusers
 from diffusers import StableDiffusionXLPipeline, DDIMScheduler, LMSDiscreteScheduler, PNDMScheduler, UniPCMultistepScheduler
 from typing import List
 import os
-import glob
 import torch
 
-def find_models(models_dir):
-    model_names = []
-    safetensor_files = glob.glob(f"{models_dir}/**/*.safetensors", recursive=True)
-    ckpt_files = glob.glob(f"{models_dir}/**/*.ckpt", recursive=True)
-
-    for file in safetensor_files:
-        if not file.endswith(".vae.safetensors"):
-            model_names.append(os.path.basename(file))
-    for file in ckpt_files:
-        if not file.endswith(".vae.ckpt"):
-            model_names.append(os.path.basename(file))
-
-    return model_names
-
-def find_vae_models(vae_dir):
-    vae_names = []
-    safetensor_files = glob.glob(f"{vae_dir}/**/*.safetensors", recursive=True)
-    ckpt_files = glob.glob(f"{vae_dir}/**/*.ckpt", recursive=True)
-
-    for file in safetensor_files:
-        vae_names.append(os.path.basename(file))
-    for file in ckpt_files:
-        vae_names.append(os.path.basename(file))
-
-    return vae_names
-
-MODELS_DIR_PATH = "models"
-VAE_DIR_PATH = "vae"
-
-MODEL_NAMES = find_models(MODELS_DIR_PATH)
-VAE_NAMES = find_vae_models(VAE_DIR_PATH)
-
-SCHEDULERS = [
-    'DDIMScheduler',
-    'LMSDiscreteScheduler',
-    'PNDMScheduler',
-    'UniPCMultistepScheduler'
-]
+from modules.sdxl_pipeline import SDXLMultiPipelineSwitchAutoDetect
+from modules.helpers import MODEL_NAMES, VAE_NAMES
+from constants.directories import MODELS_DIR_PATH, VAE_DIR_PATH
+from constants.schedulers import SCHEDULERS
 
 assert(len(MODEL_NAMES) > 0)
 assert(len(VAE_NAMES) > 0)
@@ -87,47 +53,3 @@ class Predictor(BasePredictor):
             image_paths.append(Path(img_file_path))
 
         return image_paths
-
-class SDXLMultiPipelineSwitchAutoDetect():
-
-    def __init__(self, models_dir_path, vae_dir_path, model_names, vae_names):
-        self.models_dir_path = models_dir_path
-        self.vae_dir_path = vae_dir_path
-        self.model_pipeline_dict = {model: {vae: {scheduler: None for scheduler in SCHEDULERS} for vae in vae_names} for model in model_names}
-        self.on_cuda_model = model_names[0]
-        self.on_cuda_vae = vae_names[0]
-        self.on_cuda_scheduler = SCHEDULERS[0]
-        self._load_all_models()
-        self.model_pipeline_dict[self.on_cuda_model][self.on_cuda_vae][self.on_cuda_scheduler].to("cuda")
-        
-    def get_pipeline(self, model_name, vae_name, scheduler_name):
-        pipeline = self.model_pipeline_dict.get(model_name).get(vae_name).get(scheduler_name)
-        if pipeline is None:
-            raise ValueError(f"Model '{model_name}', VAE model '{vae_name}' or scheduler '{scheduler_name}' not found or not loaded.")
-
-        if model_name != self.on_cuda_model or vae_name != self.on_cuda_vae or scheduler_name != self.on_cuda_scheduler:
-            self.model_pipeline_dict[self.on_cuda_model][self.on_cuda_vae][self.on_cuda_scheduler].to("cpu")
-            pipeline.to("cuda")
-            self.on_cuda_model = model_name
-            self.on_cuda_vae = vae_name
-            self.on_cuda_scheduler = scheduler_name
-
-        return pipeline
-
-    def _load_all_models(self):
-        for model_name in self.model_pipeline_dict.keys():
-            for vae_name in self.model_pipeline_dict[model_name]:
-                for scheduler_name in SCHEDULERS:
-                    pipeline = self._load_model(model_name, vae_name, scheduler_name)
-                    self.model_pipeline_dict[model_name][vae_name][scheduler_name] = pipeline
-
-    def _load_model(self, model_name, vae_name, scheduler_name):
-        model_path = os.path.join(self.models_dir_path, model_name)
-        vae_path = os.path.join(self.vae_dir_path, vae_name)
-        vae = diffusers.AutoencoderKL.from_single_file(vae_path, torch_dtype=torch.bfloat16)          
-        pipeline = StableDiffusionXLPipeline.from_single_file(
-            model_path, vae=vae, torch_dtype=torch.bfloat16, variant="fp16", add_watermarker=False,
-        )
-        pipeline.scheduler = globals()[scheduler_name].from_config(pipeline.scheduler.config)
-        pipeline.enable_xformers_memory_efficient_attention()
-        return pipeline
